Fix KeyStoreClientImpl::doesKeyExist to return correct result
The KeyStoreService returns NO_ERROR if the key was found,
and another response code otherwise. All of these are
mapped to non-zero values. As a result, if a key's
existence was queried it would always respond "true",
regardless if it exists or not or if there was a permissions
error.

Bug: 117993149
Test: Manual test forthcoming

Change-Id: Ib3488c6b125304fa300029e78a218f86f33d4412
Merged-In: Iffc2e155a61354f1fbffbece093b19e5cbc537fd