Use SHA256 for 32 byte keys
For NIAP certification keys need to be generated using SHA256 or
higher. Presently SHA1 is used. To satisfy this requirement,
SHA256 will be used for new keys. As the master key has recently
increased in size, the key size is used to determine if SHA1 is used
(for older keys) or SHA256.

Bug: 121272336
Test: Ran Keystore CTS tests against Walleye

Change-Id: I6099156173e04b22c6edafd9fb0e072f7201c5ee