Grant VTS tests all permissions in keystore on userdebug/eng
A VTS test for the Wifi Keystore HAL is being created. The test
is run as root and attempts various operations that directly
use the Keystore service. By default that test will not be
able to perform necessary actions to exercise the HAL code,
such as creating keys for tests.

This change will enable the root user to perform all key
operations, but only on userdebug and eng builds. In addition,
the root user will be able to perform actions on behalf of the
wifi user; this is necessary as some operations in the Wifi
Keystore HAL assume the wifi user.

Bug: 120182820
Test: atest system/hardware/interfaces/wifi/keystore/1.0/vts/
      functional/VtsHalWifiKeystoreV1_0TargetTest.cpp

Change-Id: Ic6eb5748e0e19b64a44c4bdf88a7074f7367db3d