security: remove the assert in keymaster_worker begin call
in vela platform, these two assert checks, the
"characteristics.hardwareEnforced.size()" and
"characteristics.softwareEnforced.size()" will not be 0 after
std::move operation, which is expected to be 0 after std::move
operation.
and in Android platform, these two assert checks is running with same
result, which is not 0 after std::move.
the root cause is that the underlying "hidl_vec"'s move constructor
do not clear the "&&other"'s size, so the assert at here cannot pass.
so we remove these two assert checks at here.

Signed-off-by: guoshichao <guoshichao@xiaomi.com>